split workflow related factories into their own file
