+++ opentech/apply/funds/tests/factories/__init__.py
@@ -2,8 +2,11 @@ from . import models
 from .models import * # noqa
 from . import blocks
 from .blocks import *  # noqa
+from . import workflows
+from .workflows import * # noqa
 
 __all__ = []
 
 __all__.extend(blocks.__all__)
 __all__.extend(models.__all__)
+__all__.extend(workflows.__all__)
